pyproject.toml
[project]
name = "agentic-labs"
version = "0.1.0"
description = "This is my first project to test python and agentic AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.12.3",
    "codeium-jupyter>=2.1.2",
    "datasets>=4.0.0",
    "dotenv>=0.9.9",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langchain-sandbox>=0.0.6",
    "langgraph>=0.5.0",
    "langtrace-python-sdk>=3.8.21",
    "pandas>=2.3.1",
    "pytest>=8.4.1",
]
